+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/PostgresDbTableFactory.java
@@ -18,6 +18,7 @@ package com.palantir.atlasdb.keyvalue.dbkvs.impl;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.dbkvs.PostgresDdlConfig;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.postgres.PostgresDdlTable;
+import com.palantir.atlasdb.keyvalue.dbkvs.impl.postgres.PostgresPrefixedTableNames;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.postgres.PostgresQueryFactory;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.postgres.PostgresTableInitializer;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.postgres.PostgresWriteTable;
@@ -26,9 +27,11 @@ import com.palantir.nexus.db.DBType;
 public class PostgresDbTableFactory implements DbTableFactory {
 
     private final PostgresDdlConfig config;
+    private final PostgresPrefixedTableNames prefixedTableNames;
 
     public PostgresDbTableFactory(PostgresDdlConfig config) {
         this.config = config;
+        this.prefixedTableNames = new PostgresPrefixedTableNames(config);
     }
 
     @Override
@@ -55,12 +58,17 @@ public class PostgresDbTableFactory implements DbTableFactory {
 
     @Override
     public DbWriteTable createWrite(TableReference tableRef, ConnectionSupplier conns) {
-        return new PostgresWriteTable(config, conns, tableRef, new PrefixedTableNames(config));
+        return new PostgresWriteTable(config, conns, tableRef, prefixedTableNames);
     }
 
     @Override
     public DBType getDbType() {
         return DBType.POSTGRESQL;
+    }
+
+    @Override
+    public PrefixedTableNames getPrefixedTableNames() {
+        return prefixedTableNames;
     }
 
     @Override