+++ docs/source/release_notes/release-notes.rst
@@ -254,6 +254,10 @@ v0.32.0
          - Fixed multiple scenarios where DBKVS can run into deadlocks due to unnecessary connections.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/1566>`__)
 
+    *    - |fixed|
+         - Fixed a long-held connection in Oracle table name mapping code
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/1593>`__)
+
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>
 
 =======